+++ compiler-core/src/type_.rs
@@ -993,6 +993,7 @@ impl ModuleValueConstructor {
 #[derive(Debug, Clone)]
 pub struct ModuleFunction {
     pub package: EcoString,
+    pub deprecation: Deprecation,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
@@ -1198,6 +1199,7 @@ impl ModuleInterface {
 
         Ok(ModuleFunction {
             package: self.package.clone(),
+            deprecation: value.deprecation.clone(),
         })
     }